requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.45.0
django-bootstrap-modal-forms==2.2.0
django-braces==1.14.0
django-ckeditor==6.1.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-email-verification==0.1.0
django-js-asset==1.2.2
django-paypal==1.1.1
django-redis==5.0.0
django-taggit==1.4.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
idna==2.10
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.5
